# Conflicts:
#	addon/utils/fleet-ops-options.js

+++ server/migrations/2025_08_28_054927_create_parts_table.php
@@ -30,8 +30,8 @@ return new class extends Migration {
             $table->foreignUuid('vendor_uuid')->nullable()->constrained('vendors', 'uuid')->nullOnDelete();
 
             $table->integer('quantity_on_hand')->default(0);
-            $table->decimal('unit_cost', 12, 2)->nullable();
-            $table->decimal('msrp', 12, 2)->nullable();
+            $table->bigInteger('unit_cost')->nullable(); // stored in cents
+            $table->bigInteger('msrp')->nullable();       // stored in cents
             $table->integer('reorder_point')->default(5);
             $table->integer('reorder_quantity')->default(10);
             $table->integer('max_stock_level')->nullable();
